+++ HospitalManagement/Commands/Doctors/AddDoctorCommand.cs
@@ -20,7 +20,7 @@ namespace HospitalManagement.Commands.Doctors
 
         public override void Execute(object parameter)
         {
-            _doctorsViewModel.CurrentSituation = (int)Situations.ADD;
+            _doctorsViewModel.CurrentSituation = Situations.ADD;
         }
     }
 }
