+++ HospitalManagement/Commands/Nurses/AddNurseCommand.cs
@@ -17,7 +17,7 @@ namespace HospitalManagement.Commands.Nurses
         }
         public override void Execute(object parameter)
         {
-            _nursesViewModel.CurrentSituation = (int)Situations.ADD;
+            _nursesViewModel.CurrentSituation = Situations.ADD;
         }
     }
 }
